refactor entity folder to separate folder
