+++ ormconfig.js
@@ -10,7 +10,7 @@ const {
 module.exports = {
   type: 'postgres',
   url: `postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_SERVICE}:${POSTGRES_PORT}/${POSTGRES_DB}`,
-  entities: [`${__dirname}/../routes/**/*.entity.{ts,js}`],
+  entities: [`${__dirname}/../entity/**/*.{ts,js}`],
   migrationsTableName: '_migrations',
   migrations: ['src/migration/**/*.ts'],
   cli: {
